renamed util file and improved docstring
