+++ hub/tests/storage_fixtures.py
@@ -1,4 +1,4 @@
-from hub.util.get_storage_provider import storage_provider_from_hub_path
+from hub.util.storage import storage_provider_from_hub_path
 from hub.core.storage.s3 import S3Provider
 from hub.core.storage.local import LocalProvider
 from hub.core.storage.memory import MemoryProvider
